[Fix] added new component file for components/TokenManagePage/ImageUpload.tsx and then update Media.tsx to import your new component file.
